Fix vehicle speed wronge decoding.

+++ daimler_src/host/comh/comh_fr/comh.c
@@ -1345,7 +1345,7 @@ static void encode_VehSpd_X_AR2_pdu (uint8 *buffer)
 //    st_comh_buffer_data.wheel_speed_raw_data = (u16)((((u16)(buffer[4] << 4) | (u16)(buffer[3] >> 8)) & 0x0FFF) * 10u);
     st_comh_buffer_data.wheel_speed_raw_data = 0;
     st_comh_buffer_data.wheel_speed_raw_data |= (u16)((u8)((buffer[3] & 0xF0)>>4));
-    st_comh_buffer_data.wheel_speed_raw_data |= (u16)((u8)(buffer[4]<<4));
+    st_comh_buffer_data.wheel_speed_raw_data |= (u16)(buffer[4]<<4);
     st_comh_buffer_data.wheel_speed_raw_data *= 10u;
 }
